fix for TagAtScope for empty tags  (#4200)
This PR addresses a bug of the `TagAtScope` resource, causing the
following error message when the target resource (scope) has no tags:

```
error: cannot create already existing subresource '/%2Fsubscriptions%2F{subscription_id}%2FresourceGroups%2Fresource_groupc9bbe73a/providers/Microsoft.Resources/tags/default'
```

The message indicates that the target resource (scope) cannot be adopted
by the TagAtScope resource, which would otherwise overwrite the tags of
that resource. When there are existing tags, it is appropriate to get an
error message, and the remedy is to use the `import` tag. This PR does
not change this behavior, it fixes an edge case where the target
resource has no tags.

The simplest repro is to create an RG with an empty set of tags (as
opposed to nil), and a TagAtScope for that same resource.


```yaml

name: azure-rm-yaml
runtime: yaml
description: A demonstration of tagging using TagAtScope.
resources:
  resource_group_2:
    type: azure-native:resources:ResourceGroup
    properties:
      # issue-3962 - Set the tags to an empty map to distinguish from a nil map.
      tags: {}
    options:
      ignoreChanges:
        - tags # issue-4196
  tag_2:
    type: azure-native:resources:TagAtScope
    properties:
      properties:
        tags:
          owner: tag_2
      scope: ${resource_group_2.id}
outputs:
  rg_2_name: ${resource_group_2.name}
  rg_2_tags: ${resource_group_2.tags}
```

Related to #3962