Update App to track 2025-01-01 (#4217)
This updates App to track API version 2025-01-01 instead of 2024-03-01,
accomplished by modifying `versions/v3-spec.yaml`.

The SDK was the regenerated via `make local_generate`.

Note there are breaking schema changes, but this provider is a bit of an
oddball in that we do allow breaking changes across minors.

Fixes #4205.