Bump default version for ScheduledQueryRules to 2025-01-01-preview (#4271)
Fixes: https://github.com/pulumi/pulumi-azure-native/issues/4226

As per the ticket the 2024-01-01-preview API for this resource is
deprecated and will stop working in October. As best I can tell, the
correct replacement is the 2025-01-01-preview version of this resource
as that is what the ARM template docs show. (And it's the only release
newer than the one being deprecated.)