Decouple provider from go-autorest legacy system (#4242)
## Overview

This PR is a continuation of
https://github.com/pulumi/pulumi-azure-native/pull/3630 and a part of
https://github.com/pulumi/pulumi-azure-native/issues/4047, and seeks to
fully decouple the new auth code from the old code, while leaving the
old code in place. Currently the system uses a mix of the old and new
code, necessitating that both be configured, and this leads to
incompatibilities and other limitations.

This PR does NOT remove the legacy code, which is still accessible by
setting `PULUMI_ENABLE_AZCORE_BACKEND=false`.


## Detailed Changes

### CI: use AZURE_CONFIG_DIR_FOR_TEST variable to configure the CLI
tests

To support testing of the CLI authentication mode, the CI script
prepares a temporary Azure config directory,
and then sets `AZURE_CONFIG_DIR_FOR_TEST` to its location. The test uses
this variable to set `AZURE_CONFIG_DIR` during the specific test (see
[CLI
Docs](https://learn.microsoft.com/en-us/cli/azure/azure-cli-configuration?view=azure-cli-latest#cli-configuration-file)).
During development, one may set `AZURE_CONFIG_DIR_FOR_TEST` to
`~/.azure`. The OIDC tests are of a similar nature.

### Avoiding the Legacy Auth Code Path in Provider.Configure

The Configure method was reworked to completely avoid any call to
`getAuthConfig`, which otherwise causes the legacy system to configure
itself, with undesirable effects like checking for the AZ CLI and
producing spurious errors.

See also: [prototype of total
removal](https://github.com/pulumi/pulumi-azure-native/compare/master...issue-4047)

### User Agent String

The provider composes a complex user agent string that includes the
provider version, the Azure SDK version, the "partner ID" if available,
and the value of `AZURE_HTTP_USER_AGENT` if present. See:
https://github.com/Azure/azure-sdk-for-go/issues/24025

### Subscription Auto-Detection

The `subscriptionId` provider setting is optional, and in some modes is
automatically discovered. Specifically in CLI authentication mode, the
provider is expected to pick up and use the active account (as per `az
account show`). The account information contains a token, a subscription
id, and a "cloud" identifier (usgov, china, public).

The old library was handling this discovery process, and the
azure-sdk-for-go library has no analogue. This PR ports over the
necessary code to run `az account show`. Note that the azure-sdk-for-go
library does execute the `az account get-access-token` command but is
insufficient for discovery purposes.

### New Validation Logic

The old library validates the configuration and returns validation
errors when required fields aren't specified for a given auth method.
For example, OIDc authentication requires a tenant ID and client ID.
This PR ports the validation logic over for each of the auth methods.

### New Implementation: GetClientConfig

The
[GetClientConfig](https://www.pulumi.com/registry/packages/azure-native/api-docs/authorization/getclientconfig/)
provider function was re-implemented using the new library, hewing close
to the
[terraform-provider-aruzerm](https://github.com/hashicorp/terraform-provider-azurerm/blob/572bb4f37d73f4f0d914737eaca4e5a1fd084c86/internal/clients/auth.go#L33)
implementation (MPL2).

The `GetClientConfig` function is designed to return information about
the provider's identity. The code uses the msgraph API to augment claims
extracted from the authentication token.

### New Implementation: GetClientToken

The
[GetClientToken](https://www.pulumi.com/registry/packages/azure-native/api-docs/authorization/getclienttoken/)
function is designed to return an access token for a given
_[authorization
scope](https://learn.microsoft.com/en-us/entra/identity-platform/scopes-oidc)_.
By default, the scope is set to the Azure Resource Manager scope. The
old implementation used a built-in registry of endpoints (from the
Hamilton library) to obtain the ARM scope (in addition to other
information such as the KeyVault suffix). The azure-sdk-for-go is
threadbare on this aspect and so some porting was required.

### Supported Clouds

It remains true that the provider is hardcoded to support three Azure
clouds - AzurePublic, AzureUSGovernment, and AzureChinaCloud.

## Testing

### Impact to Legacy Auth

A new integration test was added to exercise the legacy code path,
without need for a matrix test.

Closes #1464 
Closes #4223